tests: moved compat -> common and moved decorator into
