+++ gwsumm/tests/test_archive.py
@@ -29,7 +29,7 @@ from numpy import testing as nptest
 
 from gwpy.timeseries import TimeSeries
 
-from compat import unittest
+from common import unittest
 from gwsumm import (archive, data, globalv, channels)
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
